Weighted average for nodal values based on areas (to conserve averaged value in p<2 case)

+++ source/problem/Green-Naghdi/problem_data_structure/gn_data_derivative.hpp
@@ -98,7 +98,7 @@ struct Derivative {
 
     /* Averaging data */
     std::vector<uint> local_nodeID;
-    std::vector<uint> node_mult;
+    std::vector<double> node_mult;
 
     /* Transform from midpoints to vertexes */
     StatMatrix<double, SWE::n_variables, SWE::n_variables> T;
